+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - [command_cooldown] feature: now the `players with level permission 4` can bypass the `unnmaed command cooldown`.
   - Make it possible to use `/run as fake-op <player> <cmd>` to bypass the cooldown if needed.
-- [anti_build] Enhanced this module
+- [anti_build] Enhance this module
   - feature: add `enable` option for each `anti type`, making it easier to switch the functionality.
   - feature: fine-tune the event injection point, for better compatibility with other mods.
+- [command_warmup] Enhance this module
+  - feature: now the `players with level permission 4` can bypass the `command warmup`.
+  - feature: improve the event injection point, for better compatibility with other mods.